Restructure the component.
